refactor activities
