+++ app/src/main/java/es/npatarino/android/gotchallenge/view/adapters/GotHouseAdapter.java
@@ -1,0 +1,69 @@
+package es.npatarino.android.gotchallenge.view.adapters;
+
+import android.app.Activity;
+import android.support.v7.widget.RecyclerView;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+import android.widget.ImageView;
+
+import com.squareup.picasso.Picasso;
+
+import java.util.ArrayList;
+import java.util.Collection;
+import java.util.List;
+
+import es.npatarino.android.gotchallenge.R;
+import es.npatarino.android.gotchallenge.domain.GoTHouse;
+
+/**
+ * @author Antonio López.
+ */
+public class GoTHouseAdapter extends RecyclerView.Adapter<RecyclerView.ViewHolder> {
+
+    private final List<GoTHouse> gcs;
+    private Activity a;
+
+    public GoTHouseAdapter(Activity activity) {
+        this.gcs = new ArrayList<>();
+        a = activity;
+    }
+
+    public void addAll(Collection<GoTHouse> collection) {
+        for (int i = 0; i < collection.size(); i++) {
+            gcs.add((GoTHouse) collection.toArray()[i]);
+        }
+    }
+
+    @Override
+    public RecyclerView.ViewHolder onCreateViewHolder(ViewGroup parent, int viewType) {
+        return new GotCharacterViewHolder(LayoutInflater.from(parent.getContext()).inflate(R.layout.got_house_row, parent, false));
+    }
+
+    @Override
+    public void onBindViewHolder(final RecyclerView.ViewHolder holder, final int position) {
+        GotCharacterViewHolder gotCharacterViewHolder = (GotCharacterViewHolder) holder;
+        gotCharacterViewHolder.render(gcs.get(position));
+    }
+
+    @Override
+    public int getItemCount() {
+        return gcs.size();
+    }
+
+    class GotCharacterViewHolder extends RecyclerView.ViewHolder {
+
+        private static final String TAG = "GotCharacterViewHolder";
+        ImageView imp;
+
+        public GotCharacterViewHolder(View itemView) {
+            super(itemView);
+            imp = (ImageView) itemView.findViewById(R.id.ivBackground);
+        }
+
+        public void render(final GoTHouse goTHouse) {
+            Picasso.with(imp.getContext()).load(goTHouse.getHouseImageUrl()).into(imp);
+        }
+    }
+
+}
