+++ app/src/main/java/es/npatarino/android/gotchallenge/view/adapters/SectionsPagerAdapter.java
@@ -1,0 +1,43 @@
+package es.npatarino.android.gotchallenge.view.adapters;
+
+import android.support.v4.app.Fragment;
+import android.support.v4.app.FragmentManager;
+import android.support.v4.app.FragmentPagerAdapter;
+
+import es.npatarino.android.gotchallenge.view.fragment.GoTHousesListFragment;
+import es.npatarino.android.gotchallenge.view.fragment.GoTListFragment;
+
+/**
+ * @author Antonio López.
+ */
+public class SectionsPagerAdapter extends FragmentPagerAdapter {
+
+    public SectionsPagerAdapter(FragmentManager fm) {
+        super(fm);
+    }
+
+    @Override
+    public Fragment getItem(int position) {
+        if (position == 0) {
+            return new GoTListFragment();
+        } else {
+            return new GoTHousesListFragment();
+        }
+    }
+
+    @Override
+    public int getCount() {
+        return 2;
+    }
+
+    @Override
+    public CharSequence getPageTitle(int position) {
+        switch (position) {
+            case 0:
+                return "Characters";
+            case 1:
+                return "Houses";
+        }
+        return null;
+    }
+}
